Fixed examples, extracted Glove into its own module
