+++ takepod/models/impl/blcc/chain_crf.py
@@ -14,7 +14,7 @@ try:
     from keras import initializers
     from keras.engine import Layer, InputSpec
 except ImportError as ex:
-    _LOGGER.debug("Problem occured while trying to import keras. If the "
+    _LOGGER.error("Problem occured while trying to import keras. If the "
                   "library is not installed visit https://keras.io/"
                   " for more details.")
