+++ test/examples/test_model_example.py
@@ -18,7 +18,8 @@ def test_basic_batch_transform_fun():
     target_class = namedtuple("TargetBatch", ["Rating"])
     input_batch = input_class(**{"Text": np.array([[1, 3], [2, 3], [0, 4]])})
     target_batch = target_class(**{"Rating": np.array([1, 0, 4])})
-    X, y = example.basic_batch_transform_fun(input_batch, target_batch)
+    X = example.feature_extraction_fn(input_batch)
+    y = example.label_extraction_fun(target_batch)
     assert X.shape == (3, 2)
     assert y.shape == (3,)
 
@@ -34,36 +35,3 @@ def test_basic_batch_transform_fun():
 )
 def test_numericalize_pauza_rating(rating, expected_output):
     assert example.numericalize_pauza_rating(rating) == expected_output
-
-
-def test_batch_transform_fun_vectorize_avg():
-    embedding_matrix = np.array([[0, 1, 1, 4, 4],
-                                 [9, 7, 4, 3, 6],
-                                 [5, 3, 1, 2, 7],
-                                 [4, 7, 9, 3, 4],
-                                 [5, 6, 3, 1, 0]])
-    input_class = namedtuple(
-        "InputBatch", ["Text"])
-    target_class = namedtuple(
-        "TargetBatch", ["Rating"])
-    input_batch = input_class(
-        **{"Text": np.array([[1, 3], [2, 3], [0, 4]])})
-    target_batch = target_class(
-        **{"Rating": np.array([1, 0, 4])})
-
-    X, y = example.batch_transform_fun_vectorize_avg(
-        x_batch=input_batch,
-        y_batch=target_batch,
-        embedding_matrix=embedding_matrix)
-    assert X.shape == (3, 5)
-    assert y.shape == (3,)
-    expected_x = np.array([[6.5, 7, 6.5, 3, 5],
-                           [4.5, 5, 5, 2.5, 5.5],
-                           [2.5, 3.5, 2, 2.5, 2]])
-    assert np.allclose(a=X,
-                       b=expected_x,
-                       rtol=0, atol=1.e-6)
-    expected_y = [1, 0, 4]
-    assert np.allclose(a=y,
-                       b=expected_y,
-                       rtol=0, atol=1.e-6)
